Add : User에 private 누락된 부분 추가했습니다 Refactor : Account를 User와 합쳤습니다 Chore : secretKey가 외부에 노출되지 않도록 수정했습니다
