+++ src/main/java/com/steady/steadyback/domain/User.java
@@ -1,19 +1,20 @@
 package com.steady.steadyback.domain;
 
-import com.steady.steadyback.config.Account;
 import com.sun.istack.NotNull;
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
 import java.util.Collection;
+import java.util.Collections;
 
 @Entity
 @NoArgsConstructor
 @Getter
 @Table(name = "user")
-public class User {
+public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -28,15 +29,15 @@ public class User {
 
     @NotNull
     @Column(unique = true, length = 50)
-    String email;
+    private String email;
 
     @NotNull
     @Column(length = 255)
-    String password;
+    private String password;
 
     @NotNull
     @Column(length = 13)
-    String phone;
+    private String phone;
 
     @Column
     private String role;
@@ -51,13 +52,41 @@ public class User {
         this.role = "USER";
     }
 
-    public Account toAccount() {
-        return Account.builder()
-                .email(email)
-                .role(role)
-                .password(password)
-                .build();
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+
+        return Collections.singletonList(new SimpleGrantedAuthority(this.role));
     }
 
+    @Override
+    public String getPassword() {
+        return this.password;
+    }
+
+    @Override
+    public String getUsername() {
+        return this.email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
 
 }
